Make sure the results directory does get created
Using mktemp to create the results directory does require that the
parent directory already exist.

Files Changed:

	* run-tests.sh: Create the parent directory before calling mktemp.
	Add some extra error checking.

+++ run-tests.sh
@@ -100,7 +100,10 @@ done
 # ====================================================================
 
 [ ! -z "${RESULTS_DIR}" ] || error "no results directory"
-TMP_RESULTS_DIR=`mktemp -d -p ${RESULTS_DIR} "results-$(date +%F-%H%M)-XXXX"`
+mkdir -p "${RESULTS_DIR}" || error "failed to create results directory"
+TMP_RESULTS_DIR=`mktemp -d -p ${RESULTS_DIR} "results-$(date +%F-%H%M)-XXXX" 2>/dev/null`
+[ ! -z "${TMP_RESULTS_DIR}" ] || error "no run-specific results directory"
+[ -d "${TMP_RESULTS_DIR}" ] || error "no run-specific results directory found"
 rm -f ${TMP_RESULTS_DIR}/*
 RESULTS_DIR=${TMP_RESULTS_DIR}
